got check_page_size code to compile

+++ test/check.h
@@ -8,7 +8,7 @@ class Check {
 
     int check_page_size(void *ptr, size_t size, size_t page_size);
   private:
-    unsigned long long get_physaddr(void *vaddr, int *page_size);
+    unsigned long long get_physaddr(void *vaddr, size_t *page_size);
     int check_page_size(void *ptr, size_t *page_size);
 
 };
